+++ src/Akka.Persistence.EventStore/Streams/EventStoreSubscriptionDisconnectedException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Akka.Persistence.EventStore.Query;
+namespace Akka.Persistence.EventStore.Streams;
 
 public class EventStoreSubscriptionDisconnectedException(string streamName, string groupName)
     : Exception($"Subscription to stream {streamName} in group {groupName} disconnected");
